Organizing steps into individual files within the linker/steps folder
