+++ src/linker/steps/linkParent.js
@@ -1,0 +1,7 @@
+import { visit } from '../../visitors/visiting'
+
+export const linkParentStep = node => visit(node, {
+  enter(node, parent) {
+    if (parent) node.parent = parent
+  }
+})
